--- frontend/src/constants/category.js
@@ -1,3 +1,0 @@
-const CATEGORY = ['게임', '스포츠', '음악', 'e스포츠', '먹방', '여행'];
-
-export { CATEGORY };
