+++ frontend/src/constants/channelCategory.js
@@ -1,0 +1,3 @@
+const CHANNEL_CATEGORY = ['게임', '스포츠', '음악', 'e스포츠', '먹방', '여행'];
+
+export { CHANNEL_CATEGORY };
